Separate eager and lazy information
This is in preparation for having a fully eager hierarchy added.

Issue #5653
