+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/AbstractModuleComponentResolveMetadata.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
-import org.gradle.api.attributes.AttributeContainer;
 import org.gradle.api.internal.attributes.AttributeContainerInternal;
 import org.gradle.api.internal.attributes.AttributesSchemaInternal;
 import org.gradle.api.internal.attributes.EmptySchema;
@@ -44,6 +43,11 @@ import java.util.Map;
 import java.util.Set;
 
 abstract class AbstractModuleComponentResolveMetadata implements ModuleComponentResolveMetadata {
+
+    private static ImmutableAttributes extractAttributes(AbstractMutableModuleComponentResolveMetadata metadata) {
+        return ((AttributeContainerInternal) metadata.getAttributes()).asImmutable();
+    }
+
     private final ImmutableAttributesFactory attributesFactory;
     private final ModuleVersionIdentifier moduleVersionIdentifier;
     private final ModuleComponentIdentifier componentIdentifier;
@@ -53,16 +57,14 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
     @Nullable
     private final ModuleSource moduleSource;
     private final ImmutableMap<String, Configuration> configurationDefinitions;
-    private final VariantMetadataRules variantMetadataRules;
     private final ImmutableList<? extends ComponentVariant> variants;
     private final HashValue contentHash;
     private final ImmutableAttributes attributes;
 
     // Configurations are built on-demand, but only once.
-    private final Map<String, DefaultConfigurationMetadata> configurations = Maps.newHashMap();
-    private Optional<ImmutableList<? extends ConfigurationMetadata>> graphVariants;
+    private final Map<String, ConfigurationMetadata> configurations = Maps.newHashMap();
 
-    AbstractModuleComponentResolveMetadata(AbstractMutableModuleComponentResolveMetadata metadata) {
+    public AbstractModuleComponentResolveMetadata(AbstractMutableModuleComponentResolveMetadata metadata) {
         this.componentIdentifier = metadata.getId();
         this.moduleVersionIdentifier = metadata.getModuleVersionId();
         changing = metadata.isChanging();
@@ -70,34 +72,53 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
         statusScheme = metadata.getStatusScheme();
         moduleSource = metadata.getSource();
         configurationDefinitions = metadata.getConfigurationDefinitions();
-        variantMetadataRules = metadata.getVariantMetadataRules();
-        contentHash = metadata.getContentHash();
         attributesFactory = metadata.getAttributesFactory();
+        contentHash = metadata.getContentHash();
         attributes = extractAttributes(metadata);
         variants = metadata.getVariants();
     }
 
-    private static ImmutableAttributes extractAttributes(AbstractMutableModuleComponentResolveMetadata metadata) {
-        return ((AttributeContainerInternal) metadata.getAttributes()).asImmutable();
-    }
-
-
-    /**
-     * Creates a copy of the given metadata
-     */
-    AbstractModuleComponentResolveMetadata(AbstractModuleComponentResolveMetadata metadata, @Nullable ModuleSource source) {
+    public AbstractModuleComponentResolveMetadata(AbstractMutableModuleComponentResolveMetadata metadata, ImmutableList<? extends ComponentVariant> variants) {
         this.componentIdentifier = metadata.getId();
         this.moduleVersionIdentifier = metadata.getModuleVersionId();
+        changing = metadata.isChanging();
+        missing = metadata.isMissing();
+        statusScheme = metadata.getStatusScheme();
+        moduleSource = metadata.getSource();
+        configurationDefinitions = metadata.getConfigurationDefinitions();
+        attributesFactory = metadata.getAttributesFactory();
+        contentHash = metadata.getContentHash();
+        attributes = extractAttributes(metadata);
+        this.variants = variants;
+    }
+
+    public AbstractModuleComponentResolveMetadata(AbstractModuleComponentResolveMetadata metadata) {
+        this.componentIdentifier = metadata.componentIdentifier;
+        this.moduleVersionIdentifier = metadata.moduleVersionIdentifier;
         changing = metadata.changing;
         missing = metadata.missing;
         statusScheme = metadata.statusScheme;
-        moduleSource = source;
+        moduleSource = metadata.moduleSource;
         configurationDefinitions = metadata.configurationDefinitions;
-        variantMetadataRules = metadata.variantMetadataRules;
+        attributesFactory = metadata.attributesFactory;
         contentHash = metadata.contentHash;
-        attributesFactory = metadata.getAttributesFactory();
         attributes = metadata.attributes;
         variants = metadata.variants;
+    }
+
+    public AbstractModuleComponentResolveMetadata(AbstractModuleComponentResolveMetadata metadata, ModuleSource source) {
+        this.componentIdentifier = metadata.componentIdentifier;
+        this.moduleVersionIdentifier = metadata.moduleVersionIdentifier;
+        changing = metadata.changing;
+        missing = metadata.missing;
+        statusScheme = metadata.statusScheme;
+        configurationDefinitions = metadata.configurationDefinitions;
+        attributesFactory = metadata.attributesFactory;
+        contentHash = metadata.contentHash;
+        attributes = metadata.attributes;
+        variants = metadata.variants;
+
+        moduleSource = source;
     }
 
     /**
@@ -107,83 +128,11 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
     protected void copyCachedState(AbstractModuleComponentResolveMetadata metadata) {
         // Copy built-on-demand state
         configurations.putAll(metadata.configurations);
-        this.graphVariants = metadata.graphVariants;
-    }
-
-    private DefaultConfigurationMetadata populateConfigurationFromDescriptor(String name, Map<String, Configuration> configurationDefinitions, Map<String, DefaultConfigurationMetadata> configurations) {
-        DefaultConfigurationMetadata populated = configurations.get(name);
-        if (populated != null) {
-            return populated;
-        }
-
-        Configuration descriptorConfiguration = configurationDefinitions.get(name);
-        if (descriptorConfiguration == null) {
-            return null;
-        }
-
-        ImmutableList<String> hierarchy = constructHierarchy(descriptorConfiguration);
-        boolean transitive = descriptorConfiguration.isTransitive();
-        boolean visible = descriptorConfiguration.isVisible();
-        populated = createConfiguration(componentIdentifier, name, transitive, visible, hierarchy, variantMetadataRules);
-        configurations.put(name, populated);
-        return populated;
-    }
-
-    private ImmutableList<String> constructHierarchy(Configuration descriptorConfiguration) {
-        if (descriptorConfiguration.getExtendsFrom().isEmpty()) {
-            return ImmutableList.of(descriptorConfiguration.getName());
-        }
-        Set<String> accumulator = new LinkedHashSet<String>();
-        populateHierarchy(descriptorConfiguration, accumulator);
-        return ImmutableList.copyOf(accumulator);
-    }
-
-    private void populateHierarchy(Configuration metadata, Set<String> accumulator) {
-        accumulator.add(metadata.getName());
-        for (String parentName : metadata.getExtendsFrom()) {
-            Configuration parent = configurationDefinitions.get(parentName);
-            populateHierarchy(parent, accumulator);
-        }
-    }
-
-    /**
-     * Creates a {@link org.gradle.internal.component.model.ConfigurationMetadata} implementation for this component.
-     */
-    protected abstract DefaultConfigurationMetadata createConfiguration(ModuleComponentIdentifier componentId, String name, boolean transitive, boolean visible, ImmutableList<String> hierarchy, VariantMetadataRules componentMetadataRules);
-
-    /**
-     * If there are no variants defined in the metadata, but the implementation knows how to provide variants it can do that here.
-     * If it can not provide variants, absent must be returned to fall back to traditional configuration selection.
-     */
-    protected Optional<ImmutableList<? extends ConfigurationMetadata>> maybeDeriveVariants() {
-        return Optional.absent();
-    }
-
-    private Optional<ImmutableList<? extends ConfigurationMetadata>> buildVariantsForGraphTraversal(List<? extends ComponentVariant> variants) {
-        if (variants.isEmpty()) {
-            return maybeDeriveVariants();
-        }
-        ImmutableList.Builder<ConfigurationMetadata> configurations = new ImmutableList.Builder<ConfigurationMetadata>();
-        for (ComponentVariant variant : variants) {
-            configurations.add(new VariantBackedConfigurationMetadata(getId(), variant, attributes, attributesFactory, variantMetadataRules));
-        }
-        return Optional.<ImmutableList<? extends ConfigurationMetadata>>of(configurations.build());
-    }
-
-    @Nullable
-    @Override
-    public AttributesSchemaInternal getAttributesSchema() {
-        return EmptySchema.INSTANCE;
     }
 
     @Override
     public ImmutableAttributesFactory getAttributesFactory() {
         return attributesFactory;
-    }
-
-    @Override
-    public HashValue getContentHash() {
-        return contentHash;
     }
 
     @Override
@@ -194,11 +143,6 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
     @Override
     public boolean isMissing() {
         return missing;
-    }
-
-    @Override
-    public String getStatus() {
-        return attributes.getAttribute(ProjectInternal.STATUS_ATTRIBUTE);
     }
 
     @Override
@@ -222,21 +166,39 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
     }
 
     @Override
-    public ImmutableList<? extends ComponentVariant> getVariants() {
-        return variants;
-    }
-
-    @Override
-    public synchronized Optional<ImmutableList<? extends ConfigurationMetadata>> getVariantsForGraphTraversal() {
-        if (graphVariants == null) {
-            graphVariants = buildVariantsForGraphTraversal(this.variants);
-        }
-        return graphVariants;
-    }
-
-    @Override
     public String toString() {
         return componentIdentifier.getDisplayName();
+    }
+
+    @Override
+    public Set<String> getConfigurationNames() {
+        return configurationDefinitions.keySet();
+    }
+
+    @Nullable
+    @Override
+    public AttributesSchemaInternal getAttributesSchema() {
+        return EmptySchema.INSTANCE;
+    }
+
+    @Override
+    public ImmutableAttributes getAttributes() {
+        return attributes;
+    }
+
+    @Override
+    public HashValue getContentHash() {
+        return contentHash;
+    }
+
+    @Override
+    public String getStatus() {
+        return attributes.getAttribute(ProjectInternal.STATUS_ATTRIBUTE);
+    }
+
+    @Override
+    public ImmutableList<? extends ComponentVariant> getVariants() {
+        return variants;
     }
 
     @Override
@@ -245,20 +207,64 @@ abstract class AbstractModuleComponentResolveMetadata implements ModuleComponent
         return new DefaultModuleComponentArtifactMetadata(getId(), ivyArtifactName);
     }
 
-    @Override
-    public Set<String> getConfigurationNames() {
-        return configurationDefinitions.keySet();
+    protected ImmutableMap<String, Configuration> getConfigurationDefinitions() {
+        return configurationDefinitions;
     }
+
+    /**
+     * If there are no variants defined in the metadata, but the implementation knows how to provide variants it can do that here.
+     * If it can not provide variants, absent must be returned to fall back to traditional configuration selection.
+     */
+    protected Optional<ImmutableList<? extends ConfigurationMetadata>> maybeDeriveVariants() {
+        return Optional.absent();
+    }
+
 
     @Override
     public synchronized ConfigurationMetadata getConfiguration(final String name) {
         return populateConfigurationFromDescriptor(name, configurationDefinitions, configurations);
     }
 
-    @Override
-    public AttributeContainer getAttributes() {
-        return attributes;
+    private ConfigurationMetadata populateConfigurationFromDescriptor(String name, Map<String, Configuration> configurationDefinitions, Map<String, ConfigurationMetadata> configurations) {
+        ConfigurationMetadata populated = configurations.get(name);
+        if (populated != null) {
+            return populated;
+        }
+
+        Configuration descriptorConfiguration = configurationDefinitions.get(name);
+        if (descriptorConfiguration == null) {
+            return null;
+        }
+
+        ImmutableList<String> hierarchy = constructHierarchy(descriptorConfiguration);
+        boolean transitive = descriptorConfiguration.isTransitive();
+        boolean visible = descriptorConfiguration.isVisible();
+        populated = createConfiguration(getId(), name, transitive, visible, hierarchy);
+        configurations.put(name, populated);
+        return populated;
     }
+
+    private ImmutableList<String> constructHierarchy(Configuration descriptorConfiguration) {
+        if (descriptorConfiguration.getExtendsFrom().isEmpty()) {
+            return ImmutableList.of(descriptorConfiguration.getName());
+        }
+        Set<String> accumulator = new LinkedHashSet<String>();
+        populateHierarchy(descriptorConfiguration, accumulator);
+        return ImmutableList.copyOf(accumulator);
+    }
+
+    private void populateHierarchy(Configuration metadata, Set<String> accumulator) {
+        accumulator.add(metadata.getName());
+        for (String parentName : metadata.getExtendsFrom()) {
+            Configuration parent = getConfigurationDefinitions().get(parentName);
+            populateHierarchy(parent, accumulator);
+        }
+    }
+
+    /**
+     * Creates a {@link org.gradle.internal.component.model.ConfigurationMetadata} implementation for this component.
+     */
+    protected abstract ConfigurationMetadata createConfiguration(ModuleComponentIdentifier componentId, String name, boolean transitive, boolean visible, ImmutableList<String> hierarchy);
 
     @Override
     public boolean equals(Object o) {
